Fix license report task in multi-project setups (#191)
* Fix license report task in multi-project setups

Currently running the license report task in multi-project setups is frequently causing the task to fail with "FWK005 parse may not be called while parsing." This is caused by the fact that LicenseReportTask.kt uses a static (companion) parser object. If parsing is already in progress for one project, a second report task on another project will fail if it tries to parse at the same time. This can be resolved by using an instance local parser variable.

* Add reasoning to use non-static parser in license task

Ensure the issue is not re-introduced in the future, since it would be difficult to cover this with an actual test due to the race condition nature of the issue.

* Fix comment formatting with ktlintFormat

Co-authored-by: Florian Link <florian.link@relution.io>