Re-run the license task on the project configurations/dependencies changes (#213)
The idea is simple - as the license task depends on the Gradle build script itself
let's mark the build script as the task input.

A slight drawback of this approach is that the license task will re-run after every
build script change, i.e. it will be UP-TO-DATE less often. But changing the build
script should not happen that often compared to the code changes and Gradle would
need to re-configure the project anyway during the next build so re-running the license
task should be acceptable.

An alternative approach could be to declare *all* project properties that affect
the task outcome as inputs to the task.

Closes #212.